885
LayoutProvider.deleteMetadata` method should be moved into
`Maven2FileSystemProvider
