+++ strongbox-storage/strongbox-storage-api/src/main/java/org/carlspring/strongbox/providers/layout/LayoutProvider.java
@@ -14,11 +14,6 @@ import java.util.Set;
 public interface LayoutProvider<T extends ArtifactCoordinates>
 {
 
-    //TODO: we should perform delete logic in RepositoryPath and RepositoryFileSystemProvider
-    @Deprecated
-    void deleteMetadata(RepositoryPath repositoryPath)
-            throws IOException;
-
     RepositoryManagementStrategy getRepositoryManagementStrategy();
 
     @Nonnull
