+++ strongbox-storage/strongbox-storage-layout-providers/strongbox-storage-maven-layout/strongbox-storage-maven-layout-provider/src/main/java/org/carlspring/strongbox/providers/layout/Maven2LayoutProvider.java
@@ -157,139 +157,7 @@ public class Maven2LayoutProvider
     public String getAlias()
     {
         return ALIAS;
-    }
-
-    @Override
-    public void deleteMetadata(RepositoryPath artifactPath)
-    {
-        try
-        {
-            RepositoryPath artifactBasePath = artifactPath;
-            RepositoryPath artifactIdLevelPath;
-            try
-            {
-                artifactIdLevelPath = artifactBasePath.getParent();
-            }
-            catch (RepositoryRelativePathConstructionException e)
-            {
-                //it's repository root directory, so we have nothing to clean here
-                return;
-            }
-
-            if (Files.exists(artifactPath))
-            {
-
-                RepositoryFileAttributes artifactFileAttributes = Files.readAttributes(artifactPath,
-                                                                                       RepositoryFileAttributes.class);
-
-                if (!artifactFileAttributes.isDirectory())
-                {
-                    artifactBasePath = artifactBasePath.getParent();
-                    artifactIdLevelPath = artifactIdLevelPath.getParent();
-
-                    // This is at the version level
-                    try (Stream<Path> pathStream = Files.list(artifactBasePath))
-                    {
-                        Path pomPath = pathStream.filter(
-                                p -> p.getFileName().toString().endsWith(".pom")).findFirst().orElse(null);
-
-                        if (pomPath != null)
-                        {
-                            String version = ArtifactUtils.convertPathToArtifact(
-                                    RepositoryFiles.relativizePath(artifactPath))
-                                                          .getVersion();
-                            version = version == null ? pomPath.getParent().getFileName().toString() : version;
-
-                            deleteMetadataAtVersionLevel(artifactBasePath, version);
-                        }
-                    }
-
-                }
-            }
-            else
-            {
-                artifactBasePath = artifactBasePath.getParent();
-                artifactIdLevelPath = artifactIdLevelPath.getParent();
-            }
-
-            if (Files.exists(artifactIdLevelPath))
-            {
-                // This is at the artifact level
-                try (Stream<Path> pathStream = Files.list(artifactIdLevelPath))
-                {
-                    Path mavenMetadataPath = pathStream.filter(p -> p.getFileName()
-                                                                     .toString()
-                                                                     .endsWith("maven-metadata.xml"))
-                                                       .findFirst()
-                                                       .orElse(null);
-
-                    if (mavenMetadataPath != null)
-                    {
-                        String version = FilenameUtils.getName(artifactBasePath.toString());
-
-                        deleteMetadataAtArtifactLevel((RepositoryPath) mavenMetadataPath.getParent(), version);
-                    }
-                }
-            }
-        }
-        catch (IOException | XmlPullParserException e)
-        {
-            // We won't do anything in this case because it doesn't have an impact to the deletion
-            logger.error(e.getMessage(), e);
-        }
-    }
-
-    public void deleteMetadataAtVersionLevel(RepositoryPath metadataBasePath,
-                                             String version)
-            throws IOException,
-                   XmlPullParserException
-    {
-        if (ArtifactUtils.isSnapshot(version) && Files.exists(metadataBasePath))
-        {
-            Metadata metadataVersionLevel = mavenMetadataManager.readMetadata(metadataBasePath);
-            if (metadataVersionLevel != null && metadataVersionLevel.getVersioning() != null &&
-                metadataVersionLevel.getVersioning().getVersions().contains(version))
-            {
-                metadataVersionLevel.getVersioning().getVersions().remove(version);
-
-                MetadataHelper.setLastUpdated(metadataVersionLevel.getVersioning());
-
-                mavenMetadataManager.storeMetadata(metadataBasePath,
-                                                   null,
-                                                   metadataVersionLevel,
-                                                   MetadataType.SNAPSHOT_VERSION_LEVEL);
-            }
-        }
-    }
-
-    public void deleteMetadataAtArtifactLevel(RepositoryPath artifactPath,
-                                              String version)
-            throws IOException,
-                   XmlPullParserException
-    {
-        Metadata metadataVersionLevel = mavenMetadataManager.readMetadata(artifactPath);
-        if (metadataVersionLevel != null && metadataVersionLevel.getVersioning() != null)
-        {
-            metadataVersionLevel.getVersioning().getVersions().remove(version);
-
-            if (version.equals(metadataVersionLevel.getVersioning().getLatest()))
-            {
-                MetadataHelper.setLatest(metadataVersionLevel);
-            }
-
-            if (version.equals(metadataVersionLevel.getVersioning().getRelease()))
-            {
-                MetadataHelper.setRelease(metadataVersionLevel);
-            }
-
-            MetadataHelper.setLastUpdated(metadataVersionLevel.getVersioning());
-
-            mavenMetadataManager.storeMetadata(artifactPath,
-                                               null,
-                                               metadataVersionLevel,
-                                               MetadataType.ARTIFACT_ROOT_LEVEL);
-        }
-    }
+    }    
 
     @Override
     public MavenRepositoryManagementStrategy getRepositoryManagementStrategy()
