+++ strongbox-storage/strongbox-storage-layout-providers/strongbox-storage-nuget-layout-provider/src/main/java/org/carlspring/strongbox/providers/layout/NugetLayoutProvider.java
@@ -72,11 +72,6 @@ public class NugetLayoutProvider
         return path.getFileName().toString().endsWith(".nuspec");
     }
 
-    @Override
-    public void deleteMetadata(RepositoryPath repositoryPath)
-    {
-    }
-
     protected String toBase64(byte[] digest)
     {
         byte[] encoded = Base64.getEncoder()
