+++ strongbox-storage/strongbox-storage-layout-providers/strongbox-storage-p2-layout-provider/src/main/java/org/carlspring/strongbox/providers/layout/P2LayoutProvider.java
@@ -49,12 +49,6 @@ public class P2LayoutProvider
     }
 
     @Override
-    public void deleteMetadata(RepositoryPath repositoryPath)
-    {
-
-    }
-
-    @Override
     public Set<String> getDefaultArtifactCoordinateValidators()
     {
         return p2RepositoryFeatures.getDefaultArtifactCoordinateValidators();
