+++ strongbox-storage/strongbox-storage-layout-providers/strongbox-storage-raw-layout-provider/src/main/java/org/carlspring/strongbox/providers/layout/RawFileSystemProvider.java
@@ -1,8 +1,11 @@
 package org.carlspring.strongbox.providers.layout;
 
 import javax.inject.Inject;
+
+import java.io.IOException;
 import java.nio.file.spi.FileSystemProvider;
 
+import org.carlspring.strongbox.providers.io.RepositoryPath;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -28,5 +31,10 @@ public class RawFileSystemProvider extends LayoutFileSystemProvider
     {
         return layoutProvider;
     }
+
+	@Override
+	protected void deleteMetadata(RepositoryPath repositoryPath) throws IOException {
+		
+	}
 
 }
